A few Enviro Art tweaks
Tweaked a couple tiles & spots in each character level.
Also in the Hub, I tried getting rid of the excessive number of actual world tiles in the area (i.e. the huge blocks of just walls) by only outlining the walkable area and having the rest of the ground/scenery beyond the player's reach on the Doodads layer for visuals only - I was hoping this would help it stop lagging so much in the editor but I cannot tell if the load is the same tbh.
If the Hub edits cause any issues, we can just revert the changes in Hub.unity specifically and keep the rest.